+++ Backpack/Backpack.h
@@ -36,6 +36,7 @@
 #import "Label.h"
 #import "Panel.h"
 #import "Spinner.h"
+#import "StarRating.h"
 #import "Switch.h"
 #import "TappableLinkLabel.h"
 #import "TextField.h"