+++ Backpack/Calendar/Classes/BPKCalendarYearPill.h
@@ -17,6 +17,7 @@
  */
 
 #import <UIKit/UIKit.h>
+
 #import <Backpack/Font.h>
 
 @interface BPKCalendarYearPill : UIView