+++ Backpack/Dialog/Classes/BPKDialogController.m
@@ -18,13 +18,13 @@
 
 #import "BPKDialogController.h"
 
-#import "BPKDialogButtonAction.h"
-#import "BPKDialogControllerAnimator.h"
-#import "BPKDialogScrimAction.h"
-
 #import <Backpack/Button.h>
 #import <Backpack/Color.h>
 #import <Backpack/Spacing.h>
+
+#import "BPKDialogButtonAction.h"
+#import "BPKDialogControllerAnimator.h"
+#import "BPKDialogScrimAction.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -207,7 +207,8 @@ NS_ASSUME_NONNULL_BEGIN
 }
 
 - (void)_setScrimAlpha:(double)scrimAlpha {
-    self.scrimView.alpha = 0.5 * fmax(fmin(scrimAlpha, 1.0), 0.0);
+    double multiplier = UIAccessibilityIsReduceTransparencyEnabled() ? 0.9 : 0.5;
+    self.scrimView.alpha = multiplier * fmax(fmin(scrimAlpha, 1.0), 0.0);
 }
 
 #pragma mark - BPKDialogViewDelegate