+++ Backpack/Dialog/Classes/BPKDialogControllerAnimator.m
@@ -100,7 +100,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSTimeInterval)duration {
     // NOTE: Based on the duration of other transitions in UIKit.
-    return 0.35;
+    return UIAccessibilityIsReduceMotionEnabled() ? 0.0 : 0.35;
 }
 
 @end